requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
lmdb>=1.4.0
omegaconf>=2.3.0
hydra-core>=1.3.0
wandb>=0.16.0
lpips>=0.1.4
scikit-image>=0.22.0
scipy>=1.11.0
matplotlib>=3.8.0
tqdm>=4.66.0
opencv-python>=4.8.0
svgwrite>=1.4.3
pycocotools>=2.0.7
einops>=0.7.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.1.0
mypy>=1.7.0
black>=23.12.0
isort>=5.13.0

# Note: DiffVG needs to be installed separately from source
# git clone https://github.com/BachiLi/diffvg.git
# cd diffvg && python setup.py install